# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MiniShop API",
        "description": "API documentation for the MiniShop e-commerce backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current server"
        }
    ],
    "paths": {
        "/api/addresses": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "List all addresses of all users (Admin only)",
                "operationId": "dc6b190d3f07db7ed032fa181dc45150",
                "responses": {
                    "200": {
                        "description": "Addresses retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Address"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not an Admin"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Create an address for the authenticated user",
                "operationId": "6344fc023028e43a5490f1c6372bd391",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label",
                                    "full_address",
                                    "recipient_name",
                                    "recipient_phone_number"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "Home"
                                    },
                                    "full_address": {
                                        "type": "string",
                                        "example": "Jl. Merdeka No. 1, Jakarta"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Near the red gate",
                                        "nullable": true
                                    },
                                    "recipient_name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "recipient_phone_number": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "is_primary": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Address created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Address"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/addresses/me": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "List the authenticated user's own addresses",
                "operationId": "8019042f0f36cfbb7280175f0db6452e",
                "responses": {
                    "200": {
                        "description": "Addresses retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Address"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/addresses/{id}": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get a single address (owner or Admin only)",
                "operationId": "2a46bc21bf18060f0ec3a4fff2debe57",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Address"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not the owner or an Admin"
                    },
                    "404": {
                        "description": "Address not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Update an address (owner or Admin only)",
                "operationId": "b4967e4140f305290dbde0982a3fc7d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label",
                                    "full_address",
                                    "recipient_name",
                                    "recipient_phone_number"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "Home"
                                    },
                                    "full_address": {
                                        "type": "string",
                                        "example": "Jl. Merdeka No. 1, Jakarta"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Near the red gate",
                                        "nullable": true
                                    },
                                    "recipient_name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "recipient_phone_number": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "is_primary": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Address updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Address"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not the owner or an Admin"
                    },
                    "404": {
                        "description": "Address not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Delete an address (owner or Admin only)",
                "operationId": "eedadf79fef97e5ee1af0b11c8856d8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not the owner or an Admin"
                    },
                    "404": {
                        "description": "Address not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Role": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Customer"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Jane Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jane@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "081234567890"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "role": {
                        "$ref": "#/components/schemas/Role"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Address": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "label": {
                        "type": "string",
                        "example": "Home"
                    },
                    "full_address": {
                        "type": "string",
                        "example": "Jl. Merdeka No. 1, Jakarta"
                    },
                    "note": {
                        "type": "string",
                        "example": "Near the red gate",
                        "nullable": true
                    },
                    "recipient_name": {
                        "type": "string",
                        "example": "Jane Doe"
                    },
                    "recipient_phone_number": {
                        "type": "string",
                        "example": "081234567890"
                    },
                    "is_primary": {
                        "type": "boolean",
                        "example": true
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the token obtained from the login endpoint, without the \"Bearer\" prefix.",
                "bearerFormat": "Sanctum token",
                "scheme": "bearer"
            }
        }
    },
...
}